refactor: moving keyring into one folder
